requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.7
gunicorn
setuptools
django-cors-headers
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-filter==24.3
django-import-export==4.3.4
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
drf-nested-routers==0.94.1
pillow==11.1.0
python-slugify==8.0.4
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
python-dotenv~=1.0.1
psycopg2-binary
dj-database-url
whitenoise[brotli]
